Fixed an issue in prune columns test. Moved a couple common test functions to the common test file.
